### What does this PR do?

statx requests less data from the filesystem so it should in theory be faster. On some traces I've seen it make the stat() call 50% faster

Zig's std.File.getEndPos calls`fstat(fd).size` which is wasteful on linux

On Windows we check if we only asked for size and request just that instead. On macOS this becomes the same as a regular stat call.

### How did you verify your code works?

Existing tests should cover it