@jdalton 2024-03-15
I've seen stat calls add up in Node-land so this is very nice! @Jarred-Sumner you noted some perf opportunities in zig-land. Has that been raised/proposed/PR'd?