@Jarred-Sumner 2024-03-15
It's probably worth adding a `getFileSize` function if only for Windows, but this PR doesn't seem to make a meaningful difference in performance, and so far I've only seen it end up net slower than fstat().  I tested on both an Amazon Linux box on AWS and a Hetzner box with a fancy RAID + SSD configuration.

Current:
```js
   10.146 ( 0.004 ms): bun/3944291 fstat(fd: 3</home/ec2-user/bench/read-file.mjs>, statbuf: 0x7ffed1707de0) = 0
    12.903 ( 0.003 ms): bun/3944291 newfstatat(dfd: 4</proc/3944291/maps>, filename: 0x877b7f15, statbuf: 0x7ffed16def60, flag: 4096) = 0
    15.465 ( 0.002 ms): bun/3944291 fstat(fd: 8, statbuf: 0x7ffed16c8260)                                 = 0
    15.570 ( 0.003 ms): bun/3944291 fstat(fd: 8, statbuf: 0x7ffed16c7fd0)                                 = 0
    84.832 ( 0.003 ms): bun/3944291 newfstatat(dfd: 9<anon_inode:[timerfd]>, filename: 0x877b7f15, statbuf: 0x7ffed16def80, flag: 4096) = 0
    84.838 ( 0.002 ms): bun/3944291 newfstatat(dfd: 9<anon_inode:[timerfd]>, filename: 0x877b7f15, statbuf: 0x7ffed16deda0, flag: 4096) = 0
    93.468 ( 0.005 ms): bun/3944291 fstat(fd: 12</home/ec2-user/bench/read-file.mjs>, statbuf: 0x7ffed16cf510) = 0
    95.195 ( 0.002 ms): bun/3944291 fstat(fd: 14, statbuf: 0x7ffed16d2e90)                                = 0
    97.472 ( 0.004 ms): Bun Pool 0/3944308 fstat(fd: 12</home/ec2-user/bench/node_modules/mitata/src/cli.mjs>, statbuf: 0x7fc93850ed50) = 0
   142.752 ( 0.005 ms): Bun Pool 1/3944309 fstat(fd: 12, statbuf: 0x7fc933f93d50)                                = 0
   143.621 ( 0.003 ms): Bun Pool 0/3944308 fstat(fd: 12, statbuf: 0x7fc93850ed50)                                = 0
   144.100 ( 0.003 ms): Bun Pool 0/3944308 fstat(fd: 12, statbuf: 0x7fc93850ed50)                                = 0
   150.358 ( 0.004 ms): Bun Pool 0/3944308 fstat(fd: 12, statbuf: 0x7fc93850ed50)                                = 0
   151.592 ( 0.004 ms): Bun Pool 0/3944308 fstat(fd: 12, statbuf: 0x7fc93850ed50)                                = 0
   155.171 ( 0.005 ms): bun/3944291 fstat(fd: 12, statbuf: 0x7ffed16d5eb0)                                = 0
```
New:
```js
      8.141 ( 0.003 ms): bun-new/3944223 statx(dfd: 8, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7fffc2788b60) = 0
     8.234 ( 0.003 ms): bun-new/3944223 statx(dfd: 8, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7fffc27888d0) = 0
    18.100 ( 0.004 ms): bun-new/3944223 statx(dfd: 12</home/ec2-user/bench/read-file.mjs>, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7fffc278fe10) = 0
    19.484 ( 0.003 ms): bun-new/3944223 statx(dfd: 14, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7fffc2793790) = 0
    21.796 ( 0.007 ms): Bun Pool 0/3944226 statx(dfd: 12</home/ec2-user/bench/node_modules/mitata/src/cli.mjs>, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a32cafb50) = 0
    54.678 ( 0.006 ms): Bun Pool 1/3944227 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a328aeb50) = 0
    55.586 ( 0.005 ms): Bun Pool 1/3944227 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a328aeb50) = 0
    56.074 ( 0.003 ms): Bun Pool 1/3944227 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a328aeb50) = 0
    56.966 ( 0.004 ms): Bun Pool 1/3944227 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a328aeb50) = 0
    58.431 ( 0.004 ms): Bun Pool 1/3944227 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7f1a328aeb50) = 0
    62.586 ( 0.004 ms): bun-new/3944223 statx(dfd: 12, filename: 0x3ecc707b, flags: , mask: , buffer: 0x7fffc27967b0) = 0
```

Closing as it introduces complexity without having a benefit. cc @billywhizz incase you find this interesting